# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WorkLink",
        "description": "Documentación de la API REST",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refrescar token",
                "description": "Genera un nuevo token JWT",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "Token refrescado exitosamente"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/profiles": {
            "get": {
                "tags": [
                    "Freelancer Profiles"
                ],
                "summary": "Listar perfiles de freelancers",
                "description": "Retorna todos los perfiles de freelancers con su usuario y rol principal.",
                "operationId": "listFreelancerProfiles",
                "responses": {
                    "200": {
                        "description": "Perfiles obtenidos exitosamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Freelancer Profiles"
                ],
                "summary": "Crear perfil de freelancer",
                "description": "Crea un perfil de freelancer. Un freelancer crea su propio perfil; un admin puede crear el perfil indicando user_id.",
                "operationId": "createFreelancerProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "Solo requerido si el usuario autenticado es admin",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Desarrollador Full Stack con experiencia en Laravel y React"
                                    },
                                    "specialty": {
                                        "type": "string",
                                        "example": "Desarrollo Web"
                                    },
                                    "hourly_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.5
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Pachuca, Hidalgo"
                                    },
                                    "available": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "average_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Perfil creado exitosamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profiles/{id}": {
            "get": {
                "tags": [
                    "Freelancer Profiles"
                ],
                "summary": "Obtener perfil de freelancer por ID",
                "description": "Retorna los detalles del perfil, incluyendo usuario, rol, servicios, portafolio y disponibilidad.",
                "operationId": "showFreelancerProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del perfil del freelancer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil obtenido exitosamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "404": {
                        "description": "Perfil no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Freelancer Profiles"
                ],
                "summary": "Actualizar perfil de freelancer",
                "description": "Actualiza la información pública del perfil.",
                "operationId": "updateFreelancerProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del perfil a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada"
                                    },
                                    "specialty": {
                                        "type": "string",
                                        "example": "Arquitecto de Software"
                                    },
                                    "hourly_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Monterrey"
                                    },
                                    "available": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "average_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado correctamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Perfil no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Freelancer Profiles"
                ],
                "summary": "Eliminar perfil de freelancer",
                "description": "Elimina el perfil especificado.",
                "operationId": "deleteFreelancerProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del perfil a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil eliminado correctamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Perfil no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}